[BEAM-1871] Move Xml IO and related classes to new sdks/java/io/xml package.
