+++ sdks/java/core/src/main/java/org/apache/beam/sdk/io/FileBasedSource.java
@@ -60,7 +60,7 @@ import org.slf4j.LoggerFactory;
  * <p>In addition to the methods left abstract from {@code BoundedSource}, subclasses must implement
  * methods to create a sub-source and a reader for a range of a single file -
  * {@link #createForSubrangeOfFile} and {@link #createSingleFileReader}. Please refer to
- * {@link XmlSource} for an example implementation of {@code FileBasedSource}.
+ * {@link TextIO TextIO.TextSource} for an example implementation of {@code FileBasedSource}.
  *
  * @param <T> Type of records represented by the source.
  */
