+++ sdks/java/io/pom.xml
@@ -77,6 +77,7 @@
     <module>kinesis</module>
     <module>mongodb</module>
     <module>mqtt</module>
+    <module>xml</module>
   </modules>
 
   <profiles>
